Separating threads.test.js to 4 parts Separating threads.test.js to - _threadTestLibrary.js - hold the library for ease of testing - threads.test.js - to handle /threads endpoint - comments.test.js - to handle /comments endpoint - replies.test.js - to handle /replies endpoint
