+++ .eslintrc.json
@@ -14,7 +14,7 @@
     },
     "rules": {
         "indent": [2, 4, {"SwitchCase": 1}],
-        "linebreak-style": ["error", "windows"],
+        "linebreak-style": ["error", "unix"],
         "object-curly-newline": ["error", {
             "consistent": true
         }],
